ref(iroh-net): Do not add the NodeId in the magicsock span field
Since we have made MagicSock private it should always be created via
the Endpoint, unless we are internally messing around with tests or
so.  Because of this we can make sure that the Endpoint always
attaches the correct "me" field to the span.  Reducing the amount of
noise in the default log output.