fix(iroh-net): reduce noise in swarm discovery due to republish (#2685)
## Description

Since republish occurs more or less every second, for every peer in the
local network, every second we insert it in the node_map (which includes
locking), print a debug line, and for requesters, re-send a value that
was previously sent (since every time a new sender is added we
immediately send known values). This pr makes it so that we do all this
only when new values are found.

## Breaking Changes

- swarm discovery will no longer send consecutive repeated values

## Notes & open questions

the typo is unrelated to the pr but was flagged by ci

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [ ] ~~Tests if relevant.~~
- [x] All breaking changes documented.