fix(iroh-net): Fix a hot-loop when the probes time out
When all the probes time out the reportgen actor went into a hot-loop.
We slow this down by resetting the timer to be in the future again.
This way the actor loop can finish normally from the probes without
entering a hot-loop.  The actual timeout doesn't matter too much,
normally it should not happen twice as the actor should be finished by
the time another PROBES_TIMEOUT expires.