refactor(iroh): Remove custom impl of `SharedAbortingJoinHandle` (#2715)
## Description

Removes our custom implementation of `SharedAbortingJoinHandle` in favor
of a combination of
- `futures_util::future::Shared`
- `futures_util::future::MapErr` and
- `tokio_util::task::AbortOnDropHandle`
<!-- A summary of what this pull request achieves and a rough list of
changes. -->

## Breaking Changes

Should all be internal, so none.
<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

We need to turn the `JoinError` to a `String`, because `JoinError` is
`!Send`, but `String` is.
This is exactly the same behavior as what our custom
`SharedAbortingJoinHandle` was doing before.
<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- ~~[ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- ~~[ ] Tests if relevant.~~
- [x] All breaking changes documented.