fix(iroh-blobs): preserve tracing subscriber in the LocalPool (#2735)
## Description

This preserves the tracing subscriber that is installed in the current
thread when the LocalPool is created.  It then installs it into every
thread managed by the pool, ensuring that tracing output from the pool
is preserved.

## Breaking Changes

Why is this even pub?  But fine:

`iroh_blobs::util::local_pool::LocalPool` will now install the tracing
subscriber of the thread creating the pool, into each thread managed
by the pool.  Practically this should not break any code already
managing their tracing subscribers either manually inside tasks or by
setting a global subscriber before creating the pool.  But if you
really liked the behaviour of swallowing the logs on doing this it's a
breaking change.

## Notes & open questions

Maybe this should be configurable, though since this is just an
internal tool for us I think we're fine just always having this
behaviour.

Fixes #2577 
Replaces #2589

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.