fix(iroh-net): Clear the recent pong time when pong is lost
When we do not receive a pong in the timeframe we should have we
remove the last ping time and clear the best_addr.  However the path
will be re-selected from the candidates as long as the pong time is
set.

This additionally clears the pong time so that there is no longer an
indication this path works.  While the DISCO ping-pong is not
loss-protected there is already an existing check that ensures this
clear is not triggered when there are still other signs of life from
the remote node.

Note that when sending call-me-maybe the pong times are also cleared.
So while I'm not a particular fan of this this is the intended way to
use this.  Ideally the pathstate would be clearer about this data and
store this as separately.  But the PathState data needs a general
cleanup and I do not want to mix in extra state for now.