refactor(iroh-net): Remove PathState::recent_pong() (#2745)
## Description

The field itself is also directly accessible with the same visibility
as this function.  Having two ways to get to this value is a bit
confusing.  And directly using the field makes it easier when checking
users by following the references from rust-analyzer.


## Breaking Changes

None

## Notes & open questions

None

## Change checklist

- [x] Self-review.
- ~~[ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~