docs(iroh-net): Improve last_pong field docs
This is a bit confusing, at least this description is closer to the
reality.

Not mentioning PING_TIMEOUT_DURATION since that is not strictly true,
if the socket has signs of being alive but a pong is lost it will
still be considered alive and the pong timeout will not be cleared.