docs(iroh-net): Improve last_pong field docs (#2747)
## Description

This is a bit confusing, at least this description is closer to the
reality.

## Breaking Changes

None

## Notes & open questions

Not mentioning PING_TIMEOUT_DURATION since that is not strictly true,
if the socket has signs of being alive but a pong is lost it will
still be considered alive and the pong timeout will not be cleared.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
-  ~~[ ] All breaking changes documented.~~