refactor(ci): redo netsim CI (#2737)
## Description

Refactors the CI jobs for netsim. Groups it to a single parametrized
runner job which is also now a callable workflow.
Makes use of the new parallel runners and also has a quick job for just
testing perf.

## Breaking Changes

Removing the `/netsim` comment commands in favor of running it from the
actions menu.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.