feat(iroh-net): Export the Ticket trait (#2765)
## Description

This exports the Ticket trait, which is kind of needed to work with
the NodeTicket.  Since NodeTicket is already provided directly in
iroh-net it is very strange to force users to pull in iroh-base to get
the ticket trait.

Also updates the docs of tickets a little.

## Breaking Changes

None

## Notes & open questions

None

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~