refactor(iroh-net): Add tracing span context to spawned tasks (#2769)
## Description

This spawned task was missing tracing span context.  We should always
call .instrument().

## Breaking Changes

None

## Notes & open questions

I wish this wouldn't be such a manual thing.  I guess we could create
a spawn API that always does this, but even then we have to remember
to use that API and not this one.  You could probably do something
with linting for not allowing some APIs.  Anyway, not now.

## Change checklist

- [x] Self-review.
- ~~[ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~