docs(*): Document cargo features in docs (#2761)
## Description

This adds the features needed for an API to the generated docs for the
remaining workspace crates.  Only the iroh crate itself does not yet
document this.

## Breaking Changes

none

## Notes & open questions

I can for the life of me not get this to work for the iroh crate
itself...  No idea why.  Feel free to give it a try.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~