feat(iroh-net)!: Allow using a NodeId directly in connect. (#2774)
## Description

Allow using anything can be converted to a NodeAddr when connecting.
This means that we can directly connect with NodeIds, and
connect_by_node_id is no longer needed.

Also deprecate connect_by_node_id since it is no longer needed.

## Breaking Changes

- iroh-net: Endpoint::connect now takes an `impl Into<NodeAddr>` instead
of a `NodeAddr`
- iroh-net: Endpoint::connect_by_node_id is deprecated (to be removed
next release)

## Notes & open questions

Note: Doing this means that the instrumentation of connect no longer
contains the remote, which is bad I guess. Not sure how to work around
this. One way would be to have connect and connect_impl - connect_impl
is like the current connect and can be instrumented as before. Any other
way to do this?

Note: since there is a From<NodeTicket> for NodeAddr, you can also
directly dial using a ticket, which is kinda neat.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [x] All breaking changes documented.