fix(iroh-net): Keep the relay connection alive on read errors
When the connection to the relay server fails the read channel will
return a read error.  At this point the ActiveRelay actor will
passively wait until it has been asked to send something again before
it will re-establish a connection.

However if the local node has no reason to send anything to the relay
server, the connection is never re-established.  This is problematic
when the relay has remote nodes trying to send to this node.  This
doubly problematic when the connection is to the home relay: the node
just sits there thinking everything is healty and quiet, but no
traffic is reaching it.

In a node with active traffic this doesn't really show up, since a
send will be triggered quickly for an active connection and the
connection with the relay server would be re-established.

The start of the ActiveRelay run loop is the right place for this.  A
read error triggers the loop to go round, logs a read error already
and then re-estagblishes the connection.

The relay_client.is_connected() call sends a message to the relay
Client actor, and relay_client.connect() does that again.  Taking the
shortcut to only call .connect() however is not better because the
logging becomes messier.  In the common case there is one
roundrip-message to the relay Client actor and this would not improve
anyway.  The two messages for the case where a reconnect is needed
does not occur commonly.