refactor: display the socket addr and relay for a `ConnectionType::Mixed` (#2793)
## Description

Each time a `ConnectionType::Mixed` has a different SocketAddr or
RelayUrl, we correctly resend a `ConnectionType::Mixed` event to
`Endpoint::conn_type_stream`. However, when we print those changes in
something like `iroh doctor`, we only print the name of the connection
type, we don't print the socket address or relay url, so it looke like
we are sending extra `Mixed` events erroneously.

Adding more details to the `Display` of `ConnectionType`, so that it
prints the socket addrs and or relay urls.

## Change checklist

- [x] Self-review.