fix(iroh-net): DiscoveredDirectAddrs need to update the timestamp
DiscoverdDirectAddrs has a timestamp of when it was last changed,
which is used to decide if the direct addr information is recent
enough to use in call-me-maybe messages.  However when it was updated
this timestamp was not included in the comparison to see if there were
changes, so that watchers would not get updates when the direct
addresses did not change.  Unfortunately this meant the timestamp also
never updated and the DiscoveredDirectAddrs was always considered out
of date (unless there were changes in the discovered addresses).

This fixes this by putting only the DirectAddrs in the watcher.

It does a few other improvements:

- Moves all logic about DiscoveredDirectAddrs into one place.  This
  could be split out to a module now if desired.

- Tiedies up the code which creates the DirectAddrs from the
  portmapper, netcheck and the local addresses.  Again this is just a
  first pass in cleaning up this code, there is so much more that
  could be done.