fix(ci): netsim does not interact with PR comments on forks (#2777)
## Description

Seems like permissions are still messy and the netsim job fails on forks
whenever it interacts with a PR from a fork, even though it's on our own
repo and we allow it to run.

This simply drops any type of commenting on the PR from a fork.

Need to add back the required check for this job once it's merged.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.