fix(ci): make sure logs get uploaded on netsim failure (#2807)
## Description

When our netsim jobs fail, we have little visibility into the why of it.
Logs can be inspected manually on the machine but that's really not
convenient. This should make sure logs always get uploaded and attached
to the job.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.