refactor(iroh-net): Failing to bind is not a Warning log
If the bind is required the call stack already makes sure that this is
surfaced correctly.  It is in fact entirely normal if IPv6 to fail to
bind and this is already logged on INFO level.  We do not need an
additional WARN log for this.

This also doesn't need to debug-log the error, anyhow's alternate
logging mechanism is perfect for this.