refactor(iroh-net): Failing to bind is not a Warning log (#2815)
## Description

If the bind is required the call stack already makes sure that this is
surfaced correctly.  It is in fact entirely normal if IPv6 to fail to
bind and this is already logged on INFO level.  We do not need an
additional WARN log for this.

This also doesn't need to debug-log the error, anyhow's alternate
logging mechanism is perfect for this.

## Breaking Changes

None

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.