refactor(iroh-net): Debug logging should not be per packet set (#2818)
## Description

These are logged for each packet on the relay path.  That's not what
we do on debug logging.  One case is already covered by other trace
logging, so just move the "why", the other case can be a trace log.

## Breaking Changes

None

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.