feat(iroh-net): Add StaticDiscovery to provide static info to endpoints (#2825)
## Description

Add StaticProvider to provide static info to endpoints.

The long term plan is to reduce ways to interact directly with the
endpoint node map and instead provide this.

## Breaking Changes

None, this just adds stuff. Long term we want to remove
endpoint.add_node_info with this.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.