tests(iroh-net): Give this a longer timeout
On windows this sometimes times out.  But the timeout was only 5
seconds and we know our windows CI runners are slow usually.  5s
really isn't enough.  If this timeout still occurs with a 30s timeout
maybe we can do some more debugging, but it is rather likely just a
slow machine.