refactor(iroh-net)!: Rename AddrInfo to NetworkPaths
This was already the leading summary of the doc comment.

This comes from a larger refactor that was trying to remove AddrInfo
from the public API after it was observed it was only used in
discovery.  However replacing it there with it's contituent parts was
not popular with everyone, and introducing a struct just in discovery
leads to a lot of churn with nothing really being achieved.

If this stays as a struct for discovery and internal use anyway, it
might as well still be used in NodeAddr I guess.  Renaming it should
solve the confusion of what the difference between NodeAddr and
AddrInfo is as well.