feat(iroh-dns-server)!: Make http rate limit configurable (#2772)
Hello,

We are currently testing some new cloud features at
[Spacedrive](https://github.com/spacedriveapp/spacedrive), and our
implementation relies heavily on iroh. As part of this, we are deploying
our own iroh-dns-server. However, since all of our backend services
operate behind a reverse proxy, we noticed that the iroh-dns-server was
frequently hitting its rate limit because it wasn’t aware of the proxy
setup. To address this, I decided to implement a configurable rate limit
for the iroh-dns-server.

## Description

This PR adds a new entry to the `iroh-dns-server` TOML file for
configuring the HTTP rate limit. The new configuration allows for
disabling the rate limit and also supports configuring it to use the
[SmartIPKeyExtract](https://github.com/benwis/tower-governor/blob/v0.4.2/src/key_extractor.rs#L85-L119),
making it compatible with reverse proxies.

## Breaking Changes

- `iroh-dns-server`'s configuration structure now has a new field
allowing to choose the rate limiting algorithms.

## Notes & open questions

:)

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.