docs(iroh-net): Link to Endpoint in the first few paragraphs
The Endpoint is the central struct.  I think it should be linked
really early on as the most obvious link to follow.

I myself rely on this link being at the start of docs page, but
introducing this early example moved it down rather far.  So move it
back up.