tests(iroh-net): Make dht_discovery_smoke test less flaky (#2884)
## Description

This test had a hardcoded sleep(1) in there, which is a recipe for
flaky tests on CI.  Write this more defensively.


## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

I'm kind of tempted to also remove the flaky marker, but who knows.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.