fix(iroh-net): When switching to a direct path reset the mtu (#2835)
## Description

When we switch paths we reset the congestion controller, but we also
need to reset the MTU to start from 1200 again.

We have a reported case where QUIC has increased the MTU of the path
already on the relay, then switches to a direct path with an MTU of 1200
and ends up with everything blackholed. So when switching the path we
should also reset the MTU discovery to start again.

Under normal circumstances the MTU wouldn't have increased much yet
anyway, so while theoretically a slight performance decrease for most
users this will not be very noticeable.

## Breaking Changes

None.

## Notes & open questions

The functionality has already been tested and integrates well with iroh.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.