docs(iroh-dns-server): Fixup rate limit config docs (#2894)
## Description

Cargo docs was complaining about the links being incorrect.  Also the
style guide was not being followed.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

Not sure how the incorrect links made it in.

Also, when will we have a linter for the style guide...

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.