feat(iroh-net): Implement the https probe (#2903)
## Description

This implements the https probe.

There does not need to be separate IPv4 or IPv6 HTTPS probes, because we
just go with whatever the HTTPS connection works over. This is what we
need to establish a relay connection after all, this also does not care.
The only reason for this is to get some kind of latency to be able to
chose a relay server.

Luckily the https probes have been integrated in the probe plan a long
time so nothing to do there.

## Breaking Changes

The URLs served by the relay changed:
- `/relay/probe` has moved to `/ping`
- `/derp/probe` has been removed.

Unless you were manually using those URLs you will not notice these
changes, nothing in the iroh codebase ever used the changed URLs.

## Notes & open questions


Currently this uses reqwest with it's default DNS resolution. This is
not great as it is not the same DNS resolution used everywhere else. It
is however the same as done in the captive portal check. This is worth
at least looking at before merging this.

## Change checklist

- [x] Use common DNS resolver.
- [x] Update URL
- [x] Write docs for URLs of relay server.
- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.