fix(iroh-net): Make sure the rtt-actor is shutdown correctly (#2914)
## Description

When the endpoint is dropped the rtt-actor should be shut down.  This
makes sure this actually happens.

Fixes #2911

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

Also make sure the task is aborted on drop. We should always do this
anyway.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.