test(iroh-cli): increase wait on windows resumption tests (#2919)
## Description

These tests fail when windows is too slow writing and copying the files
needed. To avoid ignoring the tests, this PR increases the time we give
it to be ready on windows runs to double the original value.

## Breaking Changes

n/a

## Notes & open questions

It would still be good to try to identify if anything changed in our
code to make iroh slower on windows. This is however very low priority
so this offers a workaround.

## Change checklist

- [x] Self-review.
- [ ] ~~Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.~~
- [x] Tests if relevant.
- [ ] ~~All breaking changes documented.~~